fix package refactor
