+++ IntelliJ/src/main/java/com/nuix/innovation/enginewrapper/LicenseResolver.java
@@ -1,4 +1,4 @@
-package com.nuix.enginebaseline;
+package com.nuix.innovation.enginewrapper;
 
 import lombok.NonNull;
 import nuix.engine.Engine;
