+++ IntelliJ/src/main/java/com/nuix/innovation/enginewrapper/NuixDiagnostics.java
@@ -1,4 +1,4 @@
-package com.nuix.enginebaseline;
+package com.nuix.innovation.enginewrapper;
 
 import java.io.File;
 import java.lang.management.ManagementFactory;
