+++ IntelliJ/src/main/java/com/nuix/innovation/enginewrapper/NuixLicenseResolver.java
@@ -1,4 +1,4 @@
-package com.nuix.enginebaseline;
+package com.nuix.innovation.enginewrapper;
 
 import lombok.NonNull;
 import nuix.LicenceProperties;
