+++ IntelliJ/src/main/java/com/nuix/innovation/enginewrapper/NuixVersion.java
@@ -1,4 +1,4 @@
-package com.nuix.enginebaseline;
+package com.nuix.innovation.enginewrapper;
 
 import java.util.regex.Pattern;
