+++ IntelliJ/src/test/java/BasicTests.java
@@ -1,4 +1,4 @@
-import com.nuix.enginebaseline.NuixEngine;
+import com.nuix.innovation.enginewrapper.NuixEngine;
 import nuix.*;
 import org.junit.jupiter.api.Test;
